Update Jenkinsfile for prod

+++ Jenkinsfile
@@ -135,14 +135,23 @@ pipeline {
                         ssh -i ${SSH_KEY} -o StrictHostKeyChecking=no ${SSH_USER}@${DEPLOY_HOST} '
                             set -e
                             echo "[1] Pulizia ambiente su ${DEPLOY_HOST}..." &&
-                            mkdir -p ${REMOTE_PROJECT_PATH} &&
-                            cd ${REMOTE_PROJECT_PATH} &&
-                            docker-compose down -v --remove-orphans || true &&
-                            docker system prune -af || true &&
-                            echo "[2] Aggiorno codice..." &&
-                            git fetch origin &&
-                            git checkout ${DEPLOY_BRANCH} || git checkout -b ${DEPLOY_BRANCH} &&
-                            git reset --hard origin/${DEPLOY_BRANCH}
+                            
+                            if [ ! -d "${REMOTE_PROJECT_PATH}" ]; then
+                                echo "[!] La directory ${REMOTE_PROJECT_PATH} non esiste. Eseguo git clone..." &&
+                                cd $(dirname ${REMOTE_PROJECT_PATH}) &&
+                                git clone https://github.com/ARPA-SIMC/adriaclim-geoportal.git $(basename ${REMOTE_PROJECT_PATH}) &&
+                                cd ${REMOTE_PROJECT_PATH} &&
+                                echo "[OK] Clone completato con successo."
+                            else
+                                echo "[✓] Directory ${REMOTE_PROJECT_PATH} trovata. Procedo con aggiornamento..." &&
+                                cd ${REMOTE_PROJECT_PATH} &&
+                                docker-compose down -v --remove-orphans || true &&
+                                docker system prune -af || true &&
+                                echo "[2] Aggiorno codice..." &&
+                                git fetch origin &&
+                                git checkout ${DEPLOY_BRANCH} || git checkout -b ${DEPLOY_BRANCH} &&
+                                git reset --hard origin/${DEPLOY_BRANCH}
+                            fi
                         '
                     """
                 }
